refactor: home切换至vite
