--- client/types.ts
@@ -1,7 +1,0 @@
-import {RouteLocationNormalizedLoaded} from 'vue-router'
-import {Store} from 'vuex'
-
-export interface IAsyncDataOptions {
-  store: Store<any>
-  route: RouteLocationNormalizedLoaded
-}
