+++ packages/home/index.html
@@ -1,0 +1,22 @@
+<!DOCTYPE html>
+<html lang="cn">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <title>Freyja</title>
+    <style>
+        html, body {
+            padding: 0;
+            margin: 0;
+            height: 100%;
+        }
+        body {
+            height: calc(100% - 60px);
+        }
+    </style>
+</head>
+<body>
+    <app id="app"><!--vue-ssr-outlet--></app>
+    <script type="module" src="/src/main.ts"></script>
+</body>
+</html>
