+++ packages/home/src/assets/avatar.jpg
