+++ packages/home/src/pages/home/page.vue
