+++ packages/home/src/store/home.js
@@ -2,7 +2,7 @@
  * Created by bangbang93 on 2017/9/6.
  */
 'use strict'
-import {Fetch} from './index'
+import {Fetch} from './index.js'
 
 export default {
   namespaced: true,
